Revert "Fix type conversion warnings."
This reverts commit ff4f76444c947fcf8d87ba81a968f2bfc04e93da.

+++ src/problem.impl.cc
@@ -559,7 +559,7 @@ namespace hpp
 	}
 
 
-	inLastPathId = (Short)(planner_->getNbPaths(hppProblemId) - 1);
+	inLastPathId = planner_->getNbPaths(hppProblemId) - 1;
 	if (inLastPathId > -1) {
 	  pathLength = planner_->getPath(hppProblemId, inLastPathId)->length();
 	}
@@ -727,7 +727,7 @@ namespace hpp
 	if (CkwsRoadmapBuilderShPtr rdmbuilder =
 	    planner_->roadmapBuilderIthProblem (hppProblemId))
 	  {
-	    return Short(rdmbuilder->roadmap ()->countNodes ());
+	    return rdmbuilder->roadmap ()->countNodes ();
 	  }
 	return -1;
       }
@@ -774,7 +774,7 @@ namespace hpp
 	if (CkwsRoadmapBuilderShPtr rdmbuilder =
 	    planner_->roadmapBuilderIthProblem (hppProblemId))
 	  {
-	    return Short(rdmbuilder->roadmap ()->countEdges ());
+	    return rdmbuilder->roadmap ()->countEdges ();
 	  }
 	return -1;
       }
@@ -789,7 +789,7 @@ namespace hpp
 	if (CkwsRoadmapBuilderShPtr rdmbuilder =
 	    planner_->roadmapBuilderIthProblem (hppProblemId))
 	  {
-	    return Short(rdmbuilder->roadmap ()->countConnectedComponents ());
+	    return rdmbuilder->roadmap ()->countConnectedComponents ();
 	  }
 	return -1;
       }
